[Binding] ElementName lookups need to walk the TemplateOwner tree
So yet more testing and checking and it appears that ElementName based
bindings never use the VisualParent to find the element, they use walk
up the hierarchy using TemplateOwner.

I added a moon-unit showing that the visual tree is not used and DRT 233
has several tests which show that Bindings can find elements outside of
their own template, even if that template is itself inside a template.