[ManagedTypeInfo] Only store the Type::Kind, not the string name (again)
Implement the same fix as last time except this time round do not change the
value stored in Type::name. Instead just have an additional field which will
always store the full type name. This is the value that the parser should use
and fixes the regression where styles in the SL3 parser could not be found.