[shocker] Fix InMainThread check
Fix InMainThread check - when libshocker is loaded as a p/invoked
library inside the libmoon process there is no way to initialize
the thread detection code on the main thread, so use the fact that
tid == pid when on the main thread instead. Fixes random crashing
when running the clipboard tests in chrome (#cbhttps for instance).