[Binding] Fix up how TypeConverters are found and used.
When a binding is attached between two properties of type Foo and Bar, we
need to look up both sides for a TypeConverter. If Foo has a typeconverter
attribute we need to verify that it can handle objects of type 'Bar' (i.e. it can
do Foo->Bar conversions) before calling the forward conversion. If Foo does not
have a typeconverter, we need to check to see if 'Bar' has one and then invoke
the reverse conversion in order to do the required Foo->Bar conversion.

Fixes more subtests in drt 233 and also the new moon-unit tests.