[Loaded] Make FrameworkElement.InvokeLoaded async like the Loaded event
One subtest in DRT 233 relies on this as does the new moon-unit test. We
implement this functionality by making InvokeLoaded trigger using the 'OnEvent'
mechanism to call the managed InvokeLoaded method before the normal event is
raised. To avoid changing the existing Loaded emission behaviour (and breaking
several moonunits) we need to ignore this new OnEvent handler when checking to
see if the Loaded event has any handlers in FrameworkElement::OnIsLoadedChanged.