[PropertyPathParser] Handle Attached properties in complex paths
If we have a property like '(Foo.Bar).Baz' this needs to be parsed
as an attached Property Foo.Bar followed by a regular property Baz.
As such, we need to trim the trailing full stop after a ')' if it
exists. Otherwise we'll parse the dot as an empty property.

Required for the complex parsing tests in drt 233.