[CollectionView] Maintain the correct selected item when replacing
When we have a sorted but not grouped collection and we replace an item
in the original collection, we need to ensure that we keep the same
item selected once we handle all the change notifications even if the index
it has in the array has changed. Fixes part of 23302.