Drop the size of Moonlight::Value to 16 bytes.
The "k" field is now a combined flags & kind field.  This means that you
can't access it directly in the hopes of getting the actual kind of the
value.  You need to use GetKind() which masks thing properly.  Similarly
the managed Mono.Value has a Kind property with setter and getter that
masks this additional bit.

Also switch from a void* to a GCHandle for the boxed_valuetype, and
switch GCHandle to using a guint32 instead of a void* internally.  Given
that mono only uses 32bit ints for gchandles, this is safe, and it saves
us alignment (and 4 bytes on 64 bit machines.)

This change drops the size of Value from 24 bytes to 16 on both 32 bit
and 64 bit machines.