[Value] Ensure our flags always have a default value of 0.
We can't use a bitflag which needs to be set by default as managed structs
cannot declare a parameterless ctor to set the flag. Change the NeedsUnref
flag back to its old semantics so that a default value with the flag unset
means that "GetNeedsUnref" returns true. Fixes the recently introduced leaks

Thanks to rolf for spotting this one.