[CollectionView] Move to the next item when editing is finished
When we call CommitEdit we need to select the next item in the list
automatically. We also need to ensure that elements are inserted into
their respective groups in the same order they are in the original
collection if there is no explicit sorting applied. Fixes the new
tests and parts of 23303.