[ResourceDictionary] Fix up issues with using a System.Type key
The SL4 parser needs to store an actual System.Type as the key when
parsing xaml otherwise we can get collisions if we have an object
with a string key that matches a System.Type put in the RD by an
implicit style (or whatever).

We also have to be careful we don't remove both typeof (Button) and
the string "System.Windows.Controls.Button" when we call
ResourceDicictionary.Remove (typeof (Button)). We should only remove
the System.Type in this case.

Fixes the new moon-unit tests, a bunch of pre-existing ones which had
a MoonlightBug attribute on them and #682916.