[ResourceDictionary] Relax the check to see if the key is already there
Sometimes Silverlight barfs when you add a System.Type which matches a
string already in a ResourceDictionary and sometimes it allows it.
Remove the check completely until it can be figured out as it's breaking
some real-world sites like the one in #682916. Add a MoonlightBug to the
test which checks this and add an additional test for the 'working'
case.