[Parser] Handle x:Name and x:Key on the one element properly
Silverlight 4 RV4 allows both x:Name and x:Key to be declared on all
elements. If both are declared on an element in a ResourceDictionary it
just uses x:Key and ignores the x:Name. However for RuntimeVersion 3 and
less you are not allowed to specify both x:Name and x:Key for non-DO
objects like System.Windows.Color.