Revert "put these back.."
This reverts commit 3e2a9d0740866ea8540ff43483f069a0ae94c3ee.
We can't ifdef out the 'override' in the method declaration as it
breaks the desktop build and makes desktop moonlight completely
unusable. If we don't override these methods we will never invoke
them and will never be able to parse the generic.xaml. Declaring
the methods as 'new' is completely useless.